chat/google_meet.py
"""
Google Meet integration utilities for creating calendar events with Meet links.
"""
from django.conf import settings
from django.utils import timezone
from datetime import timedelta
import logging
import uuid

logger = logging.getLogger(__name__)

# Try to import Google API libraries (optional for development)
try:
    from google.oauth2.credentials import Credentials
    from google_auth_oauthlib.flow import Flow
    from googleapiclient.discovery import build
    from googleapiclient.errors import HttpError
    GOOGLE_APIS_AVAILABLE = True
except ImportError:
    logger.warning("Google API libraries not available - using mock Meet links")
    GOOGLE_APIS_AVAILABLE = False


def get_google_calendar_service(user_email):
    """
    Initialize Google Calendar API service with OAuth credentials.
    
    For simplicity, we'll use a service account approach or stored credentials.
    In production, you'd implement proper OAuth flow per user.
    
    For now, this uses application credentials to create events on behalf of the app.
    """
    if not GOOGLE_APIS_AVAILABLE:
        return None
        
    try:
        # For development: Use OAuth flow with stored credentials
        # In production: Implement per-user OAuth or use service account
        
        # This is a simplified version - you'll need to implement proper OAuth
        # For now, we'll create a basic credentials object
        # TODO: Implement proper OAuth flow for production
        
        flow = Flow.from_client_config(
            {
                "web": {
                    "client_id": settings.GOOGLE_CLIENT_ID,
                    "client_secret": settings.GOOGLE_CLIENT_SECRET,
                    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
                    "token_uri": "https://oauth2.googleapis.com/token",
                    "redirect_uris": [settings.GOOGLE_REDIRECT_URI],
                }
            },
            scopes=['https://www.googleapis.com/auth/calendar.events']
        )
        
        # For development, we'll return None and handle it gracefully
        # In production, implement proper OAuth flow
        return None
        
    except Exception as e:
        logger.error(f"Error initializing Google Calendar service: {e}")
        return None


def create_google_meet_event(caller_name, caller_email, receiver_name, receiver_email, call_type='video'):
    """
    Create a Google Calendar event with a Meet link.
    
    Args:
        caller_name: Name of the person initiating the call
        caller_email: Email of the caller
        receiver_name: Name of the person receiving the call
        receiver_email: Email of the receiver
        call_type: 'audio' or 'video'
    
    Returns:
        dict: {'meet_link': str, 'event_id': str} or None if failed
    """
    try:
        # For development without full OAuth: Generate a placeholder Meet link
        # In production, this would create an actual calendar event
        
        # TODO: Implement actual Google Calendar API call
        # For now, return a mock response for development
        
        logger.warning("Using mock Google Meet link - implement OAuth for production")
        
        # Generate a unique mock link for development
        mock_id = str(uuid.uuid4())[:8]
        
        return {
            'meet_link': f'https://meet.google.com/mock-{mock_id}',
            'event_id': f'mock_event_{mock_id}',
            'note': 'This is a development mock link. Implement OAuth for production.'
        }
        
        # Production implementation would look like:
        """
        service = get_google_calendar_service(caller_email)
        if not service:
            return None
        
        # Calculate event time (30 minutes from now, 1 hour duration)
        start_time = timezone.now() + timedelta(minutes=5)
        end_time = start_time + timedelta(hours=1)
        
        event = {
            'summary': f'{call_type.capitalize()} Call: {caller_name} & {receiver_name}',
            'description': f'StudyIt {call_type} call between {caller_name} and {receiver_name}',
            'start': {
                'dateTime': start_time.isoformat(),
                'timeZone': 'America/New_York',
            },
            'end': {
                'dateTime': end_time.isoformat(),
                'timeZone': 'America/New_York',
            },
            'attendees': [
                {'email': caller_email},
                {'email': receiver_email},
            ],
            'conferenceData': {
                'createRequest': {
                    'requestId': f'studyit-{uuid.uuid4()}',
                    'conferenceSolutionKey': {'type': 'hangoutsMeet'},
                }
            },
            'reminders': {
                'useDefault': False,
                'overrides': [
                    {'method': 'popup', 'minutes': 5},
                ],
            },
        }
        
        created_event = service.events().insert(
            calendarId='primary',
            body=event,
            conferenceDataVersion=1,
            sendUpdates='all'
        ).execute()
        
        meet_link = created_event.get('hangoutLink')
        event_id = created_event.get('id')
        
        return {
            'meet_link': meet_link,
            'event_id': event_id
        }
        """
        
    except Exception as e:
        logger.error(f"Error creating Google Meet event: {e}")
        return None


def delete_google_meet_event(event_id):
    """
    Delete a Google Calendar event.
    
    Args:
        event_id: The Google Calendar event ID
    
    Returns:
        bool: True if successful, False otherwise
    """
    try:
        # TODO: Implement actual deletion
        # For development, just log it
        logger.info(f"Mock deletion of event: {event_id}")
        return True
        
        # Production implementation:
        """
        service = get_google_calendar_service()
        if not service:
            return False
        
        service.events().delete(
            calendarId='primary',
            eventId=event_id,
            sendUpdates='all'
        ).execute()
        
        return True
        """
        
    except Exception as e:
        logger.error(f"Error deleting Google Meet event: {e}")
        return False
