requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2

# Google Calendar API for Meet links
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Database
# SQLite is included with Python, no additional package needed

# API & Serialization
djangorestframework==3.14.0

# CORS headers (if needed for frontend separation)
django-cors-headers==4.3.0

# Utility
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0

gunicorn
whitenoise
psycopg2-binary

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0

